backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.1
pydantic==2.7.1
python-dotenv==1.0.1
httpx==1.6.3
google-generativeai==0.1.4
